+++ frontend/src/components/HomeAdmin.jsx
@@ -2,13 +2,14 @@ import { useSelector } from 'react-redux'
 import Notification from './Notification'
 import { Table } from 'react-bootstrap'
 import { Link } from 'react-router-dom'
-import { format, formatDistanceToNow, endOfDay, isPast } from 'date-fns'
-import Badge from 'react-bootstrap/Badge';
+import { format, endOfDay } from 'date-fns'
+import DueDateBadge from './DueDateBadge'
 
 const HomeAdmin = () => {
   const user = useSelector(({ user }) => user)
   const clients = useSelector(({ clients }) =>
     clients).filter(c => c.user_id === user.id)
+  const now = endOfDay(new Date())
 
   if (!user) {
     return ('Et ole kirjautunut sisään')
@@ -40,7 +41,7 @@ const HomeAdmin = () => {
                       </td>
                       <td>{client.deadlines != '' &&
                         format(client.deadlines[0], 'dd.MM.yyyy')} {' '}
-                        <DueDateStatus client={client} />
+                        <DueDateBadge client={client} now={now} />
                       </td>
                     </tr>
                   )}
@@ -50,27 +51,6 @@ const HomeAdmin = () => {
         </div>
       }
     </div>
-  )
-}
-
-const DueDateStatus = ({ client }) => {
-  const now = endOfDay(new Date())
-  const late = isPast(client.deadlines[0])
-  
-  if (late) {
-    return (
-      <Badge bg="danger" pill>myöhässä</Badge>
-    )
-  }
-
-  let days_left = formatDistanceToNow(client.deadlines[0], now).split(" ")
-  if (days_left[0] === "about") 
-    days_left = "1 päivä"
-  else
-    days_left = days_left[0] + " päivää"
-
-  return (
-    <Badge bg="primary" pill>{days_left}</Badge>
   )
 }
